+++ testutil/key_test.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
+	codectestutil "github.com/cosmos/cosmos-sdk/codec/testutil"
 	"github.com/cosmos/cosmos-sdk/crypto/hd"
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 	"github.com/cosmos/cosmos-sdk/types"
@@ -13,7 +14,7 @@ import (
 
 func TestGenerateCoinKey(t *testing.T) {
 	t.Parallel()
-	cdc := testutil.MakeTestEncodingConfig().Codec
+	cdc := testutil.MakeTestEncodingConfig(codectestutil.CodecOptions{}).Codec
 	addr, mnemonic, err := GenerateCoinKey(hd.Secp256k1, cdc)
 	require.NoError(t, err)
 
@@ -28,7 +29,7 @@ func TestGenerateCoinKey(t *testing.T) {
 func TestGenerateSaveCoinKey(t *testing.T) {
 	t.Parallel()
 
-	encCfg := testutil.MakeTestEncodingConfig()
+	encCfg := testutil.MakeTestEncodingConfig(codectestutil.CodecOptions{})
 	kb, err := keyring.New(t.Name(), "test", t.TempDir(), nil, encCfg.Codec)
 	require.NoError(t, err)
 
@@ -53,7 +54,7 @@ func TestGenerateSaveCoinKey(t *testing.T) {
 func TestGenerateSaveCoinKeyOverwriteFlag(t *testing.T) {
 	t.Parallel()
 
-	encCfg := testutil.MakeTestEncodingConfig()
+	encCfg := testutil.MakeTestEncodingConfig(codectestutil.CodecOptions{})
 	kb, err := keyring.New(t.Name(), "test", t.TempDir(), nil, encCfg.Codec)
 	require.NoError(t, err)