+++ x/group/keeper/proposal_executor.go
@@ -2,13 +2,12 @@ package keeper
 
 import (
 	"bytes"
-	"fmt"
+	"context"
 
 	errorsmod "cosmossdk.io/errors"
 	"cosmossdk.io/x/group"
 	"cosmossdk.io/x/group/errors"
 
-	"github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
@@ -16,12 +15,13 @@ import (
 
 // doExecuteMsgs routes the messages to the registered handlers. Messages are limited to those that require no authZ or
 // by the account of group policy only. Otherwise this gives access to other peoples accounts as the sdk middlewares are bypassed
-// TODO: use context.Context and env bundler service once baseapp's MsgServiceHandler is migrated to use context.Context
-func (s Keeper) doExecuteMsgs(ctx sdk.Context, router baseapp.MessageRouter, proposal group.Proposal, groupPolicyAcc sdk.AccAddress, decisionPolicy group.DecisionPolicy) ([]sdk.Result, error) {
+func (k Keeper) doExecuteMsgs(ctx context.Context, proposal group.Proposal, groupPolicyAcc sdk.AccAddress, decisionPolicy group.DecisionPolicy) error {
+	currentTime := k.environment.HeaderService.GetHeaderInfo(ctx).Time
+
 	// Ensure it's not too early to execute the messages.
 	minExecutionDate := proposal.SubmitTime.Add(decisionPolicy.GetMinExecutionPeriod())
-	if ctx.HeaderInfo().Time.Before(minExecutionDate) {
-		return nil, errors.ErrInvalid.Wrapf("must wait until %s to execute proposal %d", minExecutionDate, proposal.Id)
+	if currentTime.Before(minExecutionDate) {
+		return errors.ErrInvalid.Wrapf("must wait until %s to execute proposal %d", minExecutionDate, proposal.Id)
 	}
 
 	// Ensure it's not too late to execute the messages.
@@ -29,37 +29,26 @@ func (s Keeper) doExecuteMsgs(ctx sdk.Context, router baseapp.MessageRouter, pro
 	// be pruned automatically, so this function should not even be called, as
 	// the proposal doesn't exist in state. For sanity check, we can still keep
 	// this simple and cheap check.
-	expiryDate := proposal.VotingPeriodEnd.Add(s.config.MaxExecutionPeriod)
-	if expiryDate.Before(ctx.HeaderInfo().Time) {
-		return nil, errors.ErrExpired.Wrapf("proposal expired on %s", expiryDate)
+	expiryDate := proposal.VotingPeriodEnd.Add(k.config.MaxExecutionPeriod)
+	if expiryDate.Before(currentTime) {
+		return errors.ErrExpired.Wrapf("proposal expired on %s", expiryDate)
 	}
 
 	msgs, err := proposal.GetMsgs()
 	if err != nil {
-		return nil, err
+		return err
 	}
 
-	results := make([]sdk.Result, len(msgs))
-	if err := ensureMsgAuthZ(msgs, groupPolicyAcc, s.cdc); err != nil {
-		return nil, err
+	if err := ensureMsgAuthZ(msgs, groupPolicyAcc, k.cdc); err != nil {
+		return err
 	}
+
 	for i, msg := range msgs {
-		handler := s.router.Handler(msg)
-		if handler == nil {
-			return nil, errorsmod.Wrapf(errors.ErrInvalid, "no message handler found for %q", sdk.MsgTypeURL(msg))
+		if _, err := k.environment.RouterService.MessageRouterService().InvokeUntyped(ctx, msg); err != nil {
+			return errorsmod.Wrapf(err, "message %s at position %d", sdk.MsgTypeURL(msg), i)
 		}
-		r, err := handler(ctx, msg)
-		if err != nil {
-			return nil, errorsmod.Wrapf(err, "message %s at position %d", sdk.MsgTypeURL(msg), i)
-		}
-		// Handler should always return non-nil sdk.Result.
-		if r == nil {
-			return nil, fmt.Errorf("got nil sdk.Result for message %q at position %d", msg, i)
-		}
-
-		results[i] = *r
 	}
-	return results, nil
+	return nil
 }
 
 // ensureMsgAuthZ checks that if a message requires signers that all of them