+++ x/tx/signing/aminojson/json_marshal.go
@@ -28,6 +28,9 @@ type EncoderOptions struct {
 	Indent string
 	// DoNotSortFields when set turns off sorting of field names.
 	DoNotSortFields bool
+	// EnumAsString when set will encode enums as strings instead of integers.
+	// Caution: Enabling this option produce different sign bytes.
+	EnumAsString bool
 	// TypeResolver is used to resolve protobuf message types by TypeURL when marshaling any packed messages.
 	TypeResolver signing.TypeResolver
 	// FileResolver is used to resolve protobuf file descriptors TypeURL when TypeResolver fails.
@@ -45,6 +48,7 @@ type Encoder struct {
 	typeResolver              protoregistry.MessageTypeResolver
 	doNotSortFields           bool
 	indent                    string
+	enumsAsString             bool
 }
 
 // NewEncoder returns a new Encoder capable of serializing protobuf messages to JSON using the Amino JSON encoding
@@ -78,6 +82,7 @@ func NewEncoder(options EncoderOptions) Encoder {
 		typeResolver:    options.TypeResolver,
 		doNotSortFields: options.DoNotSortFields,
 		indent:          options.Indent,
+		enumsAsString:   options.EnumAsString,
 	}
 	return enc
 }
@@ -189,7 +194,7 @@ func (enc Encoder) beginMarshal(msg protoreflect.Message, writer io.Writer, isAn
 		}
 	}
 
-	err := enc.marshal(protoreflect.ValueOfMessage(msg), writer)
+	err := enc.marshal(protoreflect.ValueOfMessage(msg), nil /* no field descriptor needed here */, writer)
 	if err != nil {
 		return err
 	}
@@ -204,7 +209,7 @@ func (enc Encoder) beginMarshal(msg protoreflect.Message, writer io.Writer, isAn
 	return nil
 }
 
-func (enc Encoder) marshal(value protoreflect.Value, writer io.Writer) error {
+func (enc Encoder) marshal(value protoreflect.Value, fd protoreflect.FieldDescriptor, writer io.Writer) error {
 	switch val := value.Interface().(type) {
 	case protoreflect.Message:
 		err := enc.marshalMessage(val, writer)
@@ -218,9 +223,20 @@ func (enc Encoder) marshal(value protoreflect.Value, writer io.Writer) error {
 			_, err := io.WriteString(writer, "null")
 			return err
 		}
-		return enc.marshalList(val, writer)
+		return enc.marshalList(val, fd, writer)
 
-	case string, bool, int32, uint32, []byte, protoreflect.EnumNumber:
+	case string, bool, int32, uint32, []byte:
+		return jsonMarshal(writer, val)
+
+	case protoreflect.EnumNumber:
+		if enc.enumsAsString && fd != nil {
+			desc := fd.Enum().Values().ByNumber(val)
+			if desc != nil {
+				_, err := io.WriteString(writer, fmt.Sprintf(`"%s"`, desc.Name()))
+				return err
+			}
+		}
+
 		return jsonMarshal(writer, val)
 
 	case uint64, int64:
@@ -342,7 +358,7 @@ func (enc Encoder) marshalMessage(msg protoreflect.Message, writer io.Writer) er
 				return err
 			}
 		} else {
-			err = enc.marshal(v, writer)
+			err = enc.marshal(v, f, writer)
 			if err != nil {
 				return err
 			}
@@ -371,7 +387,7 @@ func jsonMarshal(w io.Writer, v interface{}) error {
 	return err
 }
 
-func (enc Encoder) marshalList(list protoreflect.List, writer io.Writer) error {
+func (enc Encoder) marshalList(list protoreflect.List, fd protoreflect.FieldDescriptor, writer io.Writer) error {
 	n := list.Len()
 
 	_, err := io.WriteString(writer, "[")
@@ -389,7 +405,7 @@ func (enc Encoder) marshalList(list protoreflect.List, writer io.Writer) error {
 		}
 		first = false
 
-		err = enc.marshal(list.Get(i), writer)
+		err = enc.marshal(list.Get(i), fd, writer)
 		if err != nil {
 			return err
 		}